fix: call getCheckInsAndSchedules endpoint for sorted shifts
