+++ frontend/src/APIClients/VolunteerAPIClient.ts
@@ -1,5 +1,7 @@
 import { BEARER_TOKEN } from "../constants/AuthConstants";
 import {
+  CheckInWithShiftType,
+  ScheduleWithShiftType,
   UpdateVolunteerDataType,
   VolunteerResponse,
 } from "../types/VolunteerTypes";
@@ -42,7 +44,7 @@ const getVolunteerByUserId = async (
 
 const getCheckInsAndSchedules = async (
   volunteerId: string,
-): Promise<VolunteerResponse[]> => {
+): Promise<(CheckInWithShiftType | ScheduleWithShiftType)[]> => {
   try {
     const { data } = await baseAPIClient.get(
       `/volunteers/shifts/${volunteerId}`,
@@ -52,7 +54,7 @@ const getCheckInsAndSchedules = async (
     );
     return data;
   } catch (error) {
-    return error as VolunteerResponse[];
+    return error as (CheckInWithShiftType | ScheduleWithShiftType)[];
   }
 };
