+++ frontend/src/types/VolunteerTypes.ts
@@ -1,4 +1,6 @@
 import { Role, Status } from "./AuthTypes";
+import { CheckIn } from "./CheckInTypes";
+import { Schedule } from "./SchedulingTypes";
 
 export type VolunteerResponse = {
   id: string;
@@ -33,3 +35,13 @@ export type VolunteerContextAction =
       type: "SET_VOLUNTEER_STATUS";
       value: Status;
     };
+
+export enum ShiftType {
+  CHECKIN = "checkIn",
+  SCHEDULING = "scheduling",
+}
+
+export type ScheduleWithShiftType = Schedule & {
+  type: ShiftType.SCHEDULING;
+};
+export type CheckInWithShiftType = CheckIn & { type: ShiftType.CHECKIN };
